ficheros actualizados

+++ controllers/userController.js
@@ -64,6 +64,11 @@ exports.updateUser = (req, res) => {
 
 exports.deleteUser = (req, res) => {
     const {id}= req.params;
+
+    if(req.user.role !== "admin"){
+        return res.status(401).json({error: "Unauthorized access."});
+    }
+
     userModel.findByIdAndDelete(id)
     .then(user => {
         if(!user)throw new Error(`user with ID ${id} not found`);
